add another translation unit + delete useless code
