+++ demehin.maxim/P4/main.cpp
@@ -5,19 +5,19 @@ int main()
 {
   constexpr char stop = '\n';
   constexpr size_t askii_size = 128;
-  size_t max_size = 100;
+  constexpr size_t max_size = 100;
   char* str = nullptr;
-  char* temp_str = nullptr;
+  char* final_str = nullptr;
 
  try
   {
     str = new char[max_size];
-    temp_str = new char[askii_size];
+    final_str = new char[askii_size];
   }
   catch (const std::bad_alloc& e)
   {
     std::cerr << "Out of memory\n";
-    delete[] temp_str;
+    delete[] final_str;
     delete[] str;
     return 1;
   }
@@ -31,7 +31,7 @@ int main()
   {
     std::cerr << "Out of memory\n";
     delete[] str;
-    delete[] temp_str;
+    delete[] final_str;
     return 1;
   }
 
@@ -39,7 +39,7 @@ int main()
    {
      std::cerr << "Empty string\n";
      delete[] str;
-     delete[] temp_str;
+     delete[] final_str;
      return 1;
    }
 
@@ -49,36 +49,9 @@ int main()
   char ustr2[askii_size] = {};
   demehin::getUniqueSymbols(str, ustr1);
   demehin::getUniqueSymbols(base_str, ustr2);
-  size_t final_size = 0;
-  final_size = demehin::unc_sym(ustr1, ustr2, temp_str);
+  demehin::unc_sym(ustr1, ustr2, final_str);
 
-  char* final_str = nullptr;
-  try
-  {
-    final_str = new char[final_size + 1];
-  }
-  catch (std::bad_alloc & e)
-  {
-    std::cerr << "Out of memory\n";
-    delete[] final_str;
-    delete[] str;
-    return 1;
-  }
-
-  for (size_t i = 0; i < final_size; ++ i)
-  {
-    final_str[i] = temp_str[i];
-  }
-  final_str[final_size] = '\0';
-  size_t iter = 0;
-  while (final_str[iter] != '\0')
-  {
-    iter++;
-    //std::cout << final_str[iter++];
-  }
- // std::cout << '\n';
-  std::cout << temp_str << '\n';
-  delete[] temp_str;
+  std::cout << final_str << '\n';
   delete[] str;
   delete[] final_str;
 }
